[630] Separates recommendations services into separate classes | Arik
