+++ packages/core/src/ICloudRecommendationsService.ts
@@ -7,5 +7,6 @@ import { RecommendationResult } from '@cloud-carbon-footprint/common'
 export default interface ICloudRecommendationsService {
   getRecommendations(
     recommendationTarget?: string,
+    centralComputeOptimizerBucket?: string,
   ): Promise<RecommendationResult[]>
 }
